[REFACTOR] Move return_indices tests into test_sanitizer.py
